Addressed review comments
