+++ utils/firestore.js
@@ -2422,586 +2422,10 @@ const searchBoxesByLocation = async (institute, location) => {
 const cgrPackagesInTransit = async (startDate, endDate) => {
 
     // Get helper methods
-    const {getBagId, materialTypeMapping} = require('./shared');
-
-    // Dictionary to match site IDs to readable location names.
-    // Used for returning the shipped location for each package.
-    const conceptIdToSiteSpecificLocation = {
-        834825425: "HP Research Clinic",
-        752948709: "Henry Ford Main Campus",
-        570271641: "Henry Ford West Bloomfield Hospital",
-        838480167: "Henry Ford Medical Center- Fairlane",
-        706927479: "HFH Livonia Research Clinic",
-        755034888: "HFH Jackson",
-        852689772: "In-Home collection",
-        911683679: "HFH Detroit Northwest",
-        763273112: "KPCO RRL",
-        767775934: "KPGA RRL",
-        531313956: "KPHI RRL",
-        715632875: "KPNW RRL",
-        692275326: "Marshfield",
-        698283667: "Lake Hallie",
-        813701399: "Weston",
-        145191545: "Ingalls Harvey",
-        489380324: "River East",
-        120264574: "South Loop",
-        567969985: 'MF Pop-Up',
-        940329442: "Orland Park",
-        255636184: "Stevens Point",
-        813412950: "Neillsville",
-        691714762: "Rice Lake",
-        487512085: "Wisconsin Rapids",
-        983848564: "Colby Abbotsford",
-        261931804: "Minocqua",
-        665277300: "Merrill",
-        467088902: "Fargo South University",
-        589224449: "Sioux Falls Imagenetics",
-        433070901: "Sioux Falls Edith Center",
-        769594361: "Fargo Amber Valley",
-        246153539: "Bemidji Clinic",
-        777644826: "DCAM",
-        111111111: "Main Campus",
-        222222222: "Frederick",
-        574368418: "HP Park Nicollet",
-        322059622: "HFH Pop-Up",
-        127626388: "Bismarck Medical Center",
-        246137578: "Sioux Falls Sanford Center",
-        723351427: "BCC- HWC",
-        807443231: "FW All Saints",
-        288564244: "BCC- Fort Worth",
-        475614532: "BCC- Plano",
-        809370237: "BCC- Worth St",
-        856158129: "BCC- Irving",
-        436956777: "NTX Biorepository",
-        483909879: "North Garland",
-        962830330: "Waco - MacArthur",
-        397883980: "Irving",
-        117840593: "Temple CDM",
-        574104518: "Temple Roney",
-        749199085: "Temple Westfield",
-        603989030: "Killeen Main",
-        471260082: "Waco Fishpond",
-    }
-
-    // Used to match specimen tube IDs to concept IDS and vice-versa,
-    // as well as to identify valid tubes within a collection
-    const specimenCollection = {
-        numToCid: {
-            '0001': '299553921',
-            '0002': '703954371',
-            '0003': '838567176',
-            '0004': '454453939',
-            '0005': '652357376',
-            '0006': '973670172',
-            '0007': '143615646',
-            '0008': '787237543',
-            '0009': '223999569',
-            '0011': '376960806',
-            '0012': '232343615',
-            '0021': '589588440',
-            '0013': '958646668',
-            '0014': '677469051',
-            '0024': '683613884',
-            '0060': '505347689',
-        },
-        cidToNum: {
-            299553921: '0001',
-            703954371: '0002',
-            838567176: '0003',
-            454453939: '0004',
-            652357376: '0005',
-            973670172: '0006',
-            143615646: '0007',
-            787237543: '0008',
-            223999569: '0009',
-            376960806: '0011',
-            232343615: '0012',
-            589588440: '0021',
-            958646668: '0013',
-            677469051: '0014',
-            683613884: '0024',
-            505347689: '0060',
-        },
-        tubeNumList: [
-            '0001',
-            '0002',
-            '0003',
-            '0004',
-            '0005',
-            '0006',
-            '0007',
-            // '0008' and '0009' are used as containers for tubes
-            '0011',
-            '0012',
-            '0021',
-            '0013',
-            '0014',
-            '0024',
-            '0060',
-        ],
-        tubeCidList: [
-            '299553921',
-            '703954371',
-            '838567176',
-            '454453939',
-            '652357376',
-            '973670172',
-            '143615646',
-            // '787237543' and '223999569' are used as containers for tubes
-            '376960806',
-            '232343615',
-            '589588440',
-            '958646668',
-            '677469051',
-            '683613884',
-            '505347689',
-        ],
-    };
-
-    const tubeDeviationFlags = [
-        fieldMapping.tubeBrokenDeviation,
-        fieldMapping.tubeDiscardDeviation,
-        fieldMapping.tubeInsufficientVolumeDeviation,
-        fieldMapping.tubeMislabelledDeviation,
-        fieldMapping.tubeNotFoundDeviation
-    ];
-
-    // All we need is site acronym currently, but the rest is left in just in case
-    const locationConceptIDToLocationMap = {
-        834825425: {
-            siteSpecificLocation: 'HP Research Clinic',
-            siteAcronym: 'HP',
-            siteCode: '531629870',
-            siteTeam: 'HealthPartners Connect Study Team',
-            loginSiteName: 'HealthPartners Research Clinic',
-            email: 'communityresearchdepartment@healthpartners.com',
-        },
-        574368418: {
-            siteSpecificLocation: 'HP Park Nicollet',
-            siteAcronym: 'HP',
-            siteCode: '531629870',
-            siteTeam: 'HealthPartners Connect Study Team',
-            loginSiteName: 'HealthPartners Research Clinic',
-            email: 'communityresearchdepartment@healthpartners.com',
-        },
-        752948709: {
-            siteSpecificLocation: 'Henry Ford Main Campus',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        570271641: {
-            siteSpecificLocation: 'Henry Ford West Bloomfield Hospital',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        838480167: {
-            siteSpecificLocation: 'Henry Ford Medical Center- Fairlane',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        706927479: {
-            siteSpecificLocation: 'HFH Livonia Research Clinic',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        322059622: {
-            siteSpecificLocation: 'HFH Pop-Up"',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        755034888: {
-            siteSpecificLocation: 'HFH Jackson',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        911683679: {
-            siteSpecificLocation: 'HFH Detroit Northwest',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        852689772: {
-            siteSpecificLocation: 'In-Home Collection',
-            siteAcronym: 'HFHS',
-            siteCode: '548392715',
-            siteTeam: 'Henry Ford Connect Study Team',
-            loginSiteName: 'Henry Ford Health System',
-            email: 'ConnectBioHFH@hfhs.org',
-        },
-        763273112: {
-            siteSpecificLocation: 'KPCO RRL',
-            siteAcronym: 'KPCO',
-            siteCode: '125001209',
-            siteTeam: 'KPCO Connect Study Team',
-            loginSiteName: 'Kaiser Permanente Colorado',
-            email: 'Connect-Study-KPCO@kp.org',
-        },
-        767775934: {
-            siteSpecificLocation: 'KPGA RRL',
-            siteAcronym: 'KPGA',
-            siteCode: '327912200',
-            siteTeam: 'KPGA Connect Study Team',
-            loginSiteName: 'Kaiser Permanente Georgia',
-            email: 'KPGAConnectBio@kp.org',
-        },
-        531313956: {
-            siteSpecificLocation: 'KPHI RRL',
-            siteAcronym: 'KPHI',
-            siteCode: '300267574',
-            siteTeam: 'KPHI Connect Study Team',
-            loginSiteName: 'Kaiser Permanente Hawaii',
-            email: 'ConnectBioKPHI@KaiserPermanente.onmicrosoft.com'
-        },
-        715632875: {
-            siteSpecificLocation: 'KPNW RRL',
-            siteAcronym: 'KPNW',
-            siteCode: '452412599',
-            siteTeam: 'KPNW Connect Study Team',
-            loginSiteName: 'Kaiser Permanente Northwest',
-            email: 'CHR_Connect_KPNW_Bio@kpchr.org',
-        },
-        692275326: {
-            siteSpecificLocation: 'Marshfield',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        698283667:{
-            siteSpecificLocation: 'Lake Hallie',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        813701399:{
-            siteSpecificLocation: 'Weston',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        691714762:{
-            siteSpecificLocation: 'Rice Lake',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        487512085:{
-            siteSpecificLocation: 'Wisconsin Rapids',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        983848564:{
-            siteSpecificLocation: 'Colby Abbotsford',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        261931804:{
-            siteSpecificLocation: 'Minocqua',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        665277300:{
-            siteSpecificLocation: 'Merrill',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        567969985:{
-            siteSpecificLocation: 'MF Pop-Up',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        255636184:{
-            siteSpecificLocation: 'Stevens Point',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        813412950:{
-            siteSpecificLocation: 'Neillsville',
-            siteAcronym: 'MFC',
-            siteCode: '303349821',
-            siteTeam: 'Marshfield Connect Study Team',
-            loginSiteName: 'Marshfield Cancer Center',
-            email: 'connectstudy@marshfieldresearch.org'
-        },
-        589224449: {
-            siteSpecificLocation: 'Sioux Falls Imagenetics',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        127626388: {
-            siteSpecificLocation: 'Bismarck Medical Center',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        246137578: {
-            siteSpecificLocation: 'Sioux Falls Sanford Center',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        433070901: {
-            siteSpecificLocation: 'Sioux Falls Edith Center',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        769594361: {
-            siteSpecificLocation: 'Fargo Amber Valley',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        246153539: {
-            siteSpecificLocation: 'Bemidji Clinic',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        467088902: {
-            siteSpecificLocation: 'Fargo South University',
-            siteAcronym: 'SFH',
-            siteCode: '657167265',
-            siteTeam: 'Sanford Connect Study Team',
-            loginSiteName: 'Sanford Health',
-            email: 'connectstudy@sanfordhealth.org',
-        },
-        777644826: {
-            siteSpecificLocation: 'DCAM',
-            siteAcronym: 'UCM',
-            siteCode: '809703864',
-            siteTeam: 'UChicago Connect Study Team',
-            loginSiteName: 'University of Chicago Medicine',
-            email: 'connectbiospecimen@bsd.uchicago.edu',
-        },
-        145191545: {
-            siteSpecificLocation: 'Ingalls Harvey',
-            siteAcronym: 'UCM',
-            siteCode: '809703864',
-            siteTeam: 'UChicago Connect Study Team',
-            loginSiteName: 'University of Chicago Medicine',
-            email: 'connectbiospecimen@bsd.uchicago.edu',
-        },
-        489380324: {
-            siteSpecificLocation: 'River East',
-            siteAcronym: 'UCM',
-            siteCode: '809703864',
-            siteTeam: 'UChicago Connect Study Team',
-            loginSiteName: 'University of Chicago Medicine',
-            email: 'connectbiospecimen@bsd.uchicago.edu',
-        },
-        120264574: {
-            siteSpecificLocation: 'South Loop',
-            siteAcronym: 'UCM',
-            siteCode: '809703864',
-            siteTeam: 'UChicago Connect Study Team',
-            loginSiteName: 'University of Chicago Medicine',
-            email: 'connectbiospecimen@bsd.uchicago.edu',
-
-        },
-        940329442: {
-            siteSpecificLocation: 'Orland Park',
-            siteAcronym: 'UCM',
-            siteCode: '809703864',
-            siteTeam: 'UChicago Connect Study Team',
-            loginSiteName: 'University of Chicago Medicine',
-            email: 'connectbiospecimen@bsd.uchicago.edu',
-        },
-            723351427: {
-                siteSpecificLocation: 'BCC- HWC',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org',
-            },
-            807443231: {
-                siteSpecificLocation: 'FW All Saints',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org',
-            },
-            288564244: {
-                siteSpecificLocation: 'BCC- Fort Worth',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org',
-            },
-            475614532: {
-                siteSpecificLocation: 'BCC- Plano',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org',
-            },
-            809370237: {
-                siteSpecificLocation: 'BCC- Worth St',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org',
-            },
-            856158129: {
-                siteSpecificLocation: 'BCC- Irving',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org',
-            },
-            436956777: {
-                siteSpecificLocation: 'NTX Biorepository',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            483909879: {
-                siteSpecificLocation: 'North Garland',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            962830330: {
-                siteSpecificLocation: 'Waco - MacArthur',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            397883980: {
-                siteSpecificLocation: 'Irving',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            117840593: {
-                siteSpecificLocation: 'Temple CDM',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            574104518: {
-                siteSpecificLocation: 'Temple Roney',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            749199085: {
-                siteSpecificLocation: 'Temple Westfield',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            603989030: {
-                siteSpecificLocation: 'Killeen Main',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            471260082: {
-                siteSpecificLocation: 'Waco Fishpond',
-                siteAcronym: 'BSWH',
-                siteCode: '472940358',
-                siteTeam: 'BSWH Connect Study Team',
-                loginSiteName: 'Baylor Scott & White Health',
-                email: 'connectbiospecimen@BSWHealth.org'
-            },
-            111111111: {
-                siteSpecificLocation: 'Main Campus',
-                siteAcronym: 'NIH',
-                siteCode: '13',
-                siteTeam: "NIH Connect Study Team",
-                loginSiteName: 'National Cancer Institute',
-                email: "connectstudytest@email.com",
-            },
-            222222222: { 
-                siteSpecificLocation: 'Frederick',
-                siteAcronym: 'NIH',
-                siteCode: '13',
-                siteTeam: "NIH Connect Study Team",
-                loginSiteName: 'National Cancer Institute',
-                email: "connectstudytest@email.com",
-            },
-    };
+    const {
+        getBagId, materialTypeMapping, 
+        conceptIdToSiteSpecificLocation, specimenCollection, tubeDeviationFlags, locationConceptIDToLocationMap
+    } = require('./shared');
 
     const bagConceptIdList = Object.values(fieldMapping.bagContainerCids);
